requirements.txt
pytorch
torchvision
os